refactor: centralize type declarations
